fix: remove userId from the configurationAndroid class (#448)
* fix: remove userId from ConfigurationAndroid and ConfigurationAndroidBuilder

UserId should not be part of the sdk configuration. It can only be modified by the identify event.

* fix: remove the code to update the userId in configuration object

* refactor: use `androidStorage` to fetch the userId

* test: fix failing test case in FillDefaultsPluginTest

Use real storage object so that userId could be fetched

* refactor: directly set the userId in storage

Analytics.setUserId extension function was required for two purpose: First is to store the userId in storage and second is to update the userId configuration. Now we no longer require the latter part we can simplify the logic.

* chore: remove setting of userId from v2Cache

As we are no longer updating the userId in configuration object, we don't need to basically fetch the userId from the storage and again store it in storage.

* refactor: create an extension function for empty string and use that